# openapi.yaml
openapi: 3.0.3
info:
  title: Anna's Archive API
  description: |
    A RESTful API to search and retrieve book metadata from Anna's Archive.
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:3000
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
          format: date-time
        rateLimitBookDetail:
          type: integer
          description: Max requests per minute for /api/book/:md5 (book detail). -1 = unlimited, 0 = blocked.
          default: -1
        rateLimitBookDownload:
          type: integer
          description: Max requests per minute for /api/book/:md5/download. -1 = unlimited, 0 = blocked.
          default: -1
        rateLimitBookRelated:
          type: integer
          description: Max requests per minute for /api/book/:md5/related. -1 = unlimited, 0 = blocked.
          default: -1
        rateLimitSearch:
          type: integer
          description: Max requests per minute for /api/search. -1 = unlimited, 0 = blocked.
          default: -1
    ApiUsageRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        endpoint:
          type: string
        method:
          type: string
        ip:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
...
paths:
  /api/auth/login:
    post:
      tags: [Auth]
      summary: User login
      description: Exchange credentials for a JWT token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      description: Admin only. Create a new user account.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
                role:
                  type: string
                  enum: [user, admin]
                  default: user
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/User'
        '403':
          description: Forbidden (Admin role required)
  /api/auth/users:
    get:
      tags: [Auth]
      summary: List all users
      description: Admin only. Retrieve a list of all registered users.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
  /api/auth/users/{id}:
    put:
      tags: [Auth]
      summary: Update a user
      description: Admin only. Update username, password, or role of an existing user.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                role:
                  type: string
                  enum: [user, admin]
                rateLimitBookDetail:
                  type: integer
                  description: Max requests per minute for /api/book/:md5. -1 = unlimited, 0 = blocked.
                rateLimitBookDownload:
                  type: integer
                  description: Max requests per minute for /api/book/:md5/download. -1 = unlimited, 0 = blocked.
                rateLimitBookRelated:
                  type: integer
                  description: Max requests per minute for /api/book/:md5/related. -1 = unlimited, 0 = blocked.
                rateLimitSearch:
                  type: integer
                  description: Max requests per minute for /api/search. -1 = unlimited, 0 = blocked.
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/User'
        '404':
          description: User not found
  /api/auth/me:
    get:
      tags: [Auth]
      summary: View own account
      description: Retrieve your own account details and current rate limit usage progress.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Account details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/User'
                  rateLimits:
                    type: object
                    properties:
                      bookDetail:
                        type: object
                        properties:
                          limit:
                            type: integer
                            nullable: true
                          current:
                            type: integer
                      bookDownload:
                        type: object
                        properties:
                          limit:
                            type: integer
                            nullable: true
                          current:
                            type: integer
                      bookRelated:
                        type: object
                        properties:
                          limit:
                            type: integer
                            nullable: true
                          current:
                            type: integer
                      search:
                        type: object
                        properties:
                          limit:
                            type: integer
                            nullable: true
                          current:
                            type: integer
        '401':
          description: Unauthorized
  /api/auth/users/{id}/usage:
    get:
      tags: [Auth]
      summary: View user API usage
      description: Admin only. Retrieve paginated API usage records and current rate limit usage for a specific user.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          description: Max records to return (max 500)
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 500
        - name: offset
          in: query
          description: Number of records to skip
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Usage records
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/User'
                  rateLimits:
                    type: object
                    properties:
                      bookDetail:
                        type: object
                        properties:
                          limit:
                            type: integer
                            nullable: true
                          current:
                            type: integer
                      bookDownload:
                        type: object
                        properties:
                          limit:
                            type: integer
                            nullable: true
                          current:
                            type: integer
                      bookRelated:
                        type: object
                        properties:
                          limit:
                            type: integer
                            nullable: true
                          current:
                            type: integer
                      search:
                        type: object
                        properties:
                          limit:
                            type: integer
                            nullable: true
                          current:
                            type: integer
                  usage:
                    type: object
                    properties:
                      total:
                        type: integer
                      limit:
                        type: integer
                      offset:
                        type: integer
                      records:
                        type: array
                        items:
                          $ref: '#/components/schemas/ApiUsageRecord'
        '403':
          description: Forbidden (Admin role required)
        '404':
          description: User not found
...